# Conflicts:
#	_std/types.dm
#	code/datums/controllers/explosion_controls.dm
#	code/modules/worldgen/prefabs.dm
#	code/turf/turf_autoalign.dm

+++ code/datums/buildmode.dm
@@ -1,19 +1,24 @@
 ABSTRACT_TYPE(/datum/buildmode)
 /datum/buildmode
-	var/list/extra_buttons = list()
+	var/tmp/list/extra_buttons = list()
 	var/hotkey_number = null
 	var/name = "You shouldn't see me."
-	var/desc = "<span class='alert'>Someone is a lazy bum.</span>"
+	var/tmp/desc = SPAN_ALERT("Someone is a lazy bum.")
 	var/datum/buildmode_holder/holder = null
 	var/icon_state = null
 	var/admin_level = LEVEL_BABBY // restricts certain things to certain ranks
-	var/atom/movable/screen/buildmode/hotkey/hotkey_button = null
+	var/tmp/atom/movable/screen/buildmode/hotkey/hotkey_button = null
 	var/button_mode_text
 
 	New(var/datum/buildmode_holder/H)
 		..()
 		holder = H
 		update_button_text()
+
+	Read(F)
+		. = ..()
+		if(!isnull(src.hotkey_number))
+			on_add_to_hotkey_bar()
 
 	proc/copy()
 		var/datum/buildmode/new_mode = semi_deep_copy(src, copy_flags=COPY_SHALLOW_EXCEPT_FOR_LISTS)
@@ -50,6 +55,7 @@ ABSTRACT_TYPE(/datum/buildmode)
 		src.hotkey_button.desc = src.desc
 		src.hotkey_button.icon_state = src.icon_state
 		src.hotkey_button.maptext = "<span class='ol pixel'>[src.hotkey_number % 10]</span>"
+		src.hotkey_button.screen_loc = "NORTH,WEST+[hotkey_number + 5]"
 
 	// Called when exiting buildmode
 	proc/paused()
@@ -77,35 +83,46 @@ ABSTRACT_TYPE(/datum/buildmode)
 	proc/click_right(atom/object, var/ctrl, var/alt, var/shift)
 
 /datum/buildmode_holder
-	var/client/owner
-	var/datum/buildmode/mode
-	var/list/modes_cache = list()
+	var/tmp/client/owner
+	var/tmp/datum/buildmode/mode
+	var/tmp/list/modes_cache = list()
 	var/list/hotkey_bar
 
-	var/is_active = 0
-	var/dir = SOUTH
+	var/tmp/is_active = 0
+	var/tmp/dir = SOUTH
 
 	New(var/client/C)
 		..()
-		hotkey_bar = list()
-		hotkey_bar.len = 10
+		if(isnull(hotkey_bar))
+			hotkey_bar = list()
+			hotkey_bar.len = 10
 
-		owner = C
 		button_dir = new(null, src)
 		button_help = new(null, src)
 		button_mode = new(null, src)
 		button_quit = new(null, src)
 
-		for (var/T in concrete_typesof(/datum/buildmode))
-			var/datum/buildmode/M = new T(src)
-			if ((!owner.holder && M.admin_level > LEVEL_BABBY) || M.admin_level > owner.holder.level)
-				DEBUG_MESSAGE("[key_name(owner)] is too low rank to have buildmode [M.name] ([M.type]) and the buildmode is being disposed (min level is [level_to_rank(M.admin_level)] and [owner.ckey] is [owner.holder ? level_to_rank(owner.holder.level) : "not an admin"])")
-				qdel(M)
-				continue
-			if (!mode || istype(M, /datum/buildmode/spawn_single))
-				select_mode(M)
-			modes_cache += M.name
-			modes_cache[M.name] = M
+		if(C)
+			set_client(C)
+
+	Read(F)
+		. = ..()
+		if(!isnull(src.hotkey_bar[1]))
+			select_mode(src.hotkey_bar[1])
+
+	proc/set_client(client/C)
+		owner = C
+		if(!length(modes_cache))
+			for (var/T in concrete_typesof(/datum/buildmode))
+				var/datum/buildmode/M = new T(src)
+				if ((!owner.holder && M.admin_level > LEVEL_BABBY) || M.admin_level > owner.holder.level)
+					DEBUG_MESSAGE("[key_name(owner)] is too low rank to have buildmode [M.name] ([M.type]) and the buildmode is being disposed (min level is [level_to_rank(M.admin_level)] and [owner.ckey] is [owner.holder ? level_to_rank(owner.holder.level) : "not an admin"])")
+					qdel(M)
+					continue
+				if (!mode || istype(M, /datum/buildmode/spawn_single))
+					select_mode(M)
+				modes_cache += M.name
+				modes_cache[M.name] = M
 
 	proc/select_mode(var/datum/buildmode/M)
 		if (mode)
@@ -135,8 +152,8 @@ ABSTRACT_TYPE(/datum/buildmode)
 	proc/update_mode_icon()
 		button_mode.icon_state = mode.icon_state
 		button_mode.maptext_width = 256
-		button_mode.maptext_height = 64
-		button_mode.maptext_y = -64
+		button_mode.maptext_height = 96
+		button_mode.maptext_y = -96
 		button_mode.maptext_x = -62
 
 	proc/build_click(atom/target, location, control, list/params)
@@ -187,7 +204,6 @@ ABSTRACT_TYPE(/datum/buildmode)
 			src.hotkey_bar[index] = null
 			mode.hotkey_number = null
 			return FALSE
-		mode.hotkey_button.screen_loc = "NORTH,WEST+[index + 5]"
 		owner.screen += mode.hotkey_button
 		src.select_mode(mode)
 		return TRUE
@@ -206,21 +222,40 @@ ABSTRACT_TYPE(/datum/buildmode)
 		return TRUE
 
 	proc/display_help()
-		boutput(usr, "<span class='notice'>[mode.desc]</span>")
+		boutput(usr, SPAN_NOTICE("[mode.desc]"))
 
 	// You shouldn't actually interact with these anymore.
-	var/atom/movable/screen/buildmode/builddir/button_dir
-	var/atom/movable/screen/buildmode/buildhelp/button_help
-	var/atom/movable/screen/buildmode/buildmode/button_mode
-	var/atom/movable/screen/buildmode/buildquit/button_quit
+	var/tmp/atom/movable/screen/buildmode/builddir/button_dir
+	var/tmp/atom/movable/screen/buildmode/buildhelp/button_help
+	var/tmp/atom/movable/screen/buildmode/buildmode/button_mode
+	var/tmp/atom/movable/screen/buildmode/buildquit/button_quit
+
+/client/proc/resetbuildmode()
+	set name = "Reset Build Mode"
+	set desc = "If your build mode save got screwed up use this to reset it!"
+	SET_ADMIN_CAT(ADMIN_CAT_SELF)
+	ADMIN_ONLY
+	SHOW_VERB_DESC
+
+	if(src.buildmode?.is_active)
+		src.togglebuildmode()
+	qdel(src.buildmode)
+	src.buildmode = new(src)
+	src.player.buildmode = src.buildmode
+	src.player.cloudSaves.deleteData("buildmode")
+	src.togglebuildmode()
 
 /client/proc/togglebuildmode()
 	set name = "Build Mode"
 	set desc = "Toggle build Mode on/off."
 	SET_ADMIN_CAT(ADMIN_CAT_SELF)
+	ADMIN_ONLY
+	SHOW_VERB_DESC
 
 	if(!src.buildmode)
-		src.buildmode = new /datum/buildmode_holder(src)
+		src.buildmode = src.player.get_buildmode()
+		src.buildmode.set_client(src)
+	logTheThing(LOG_ADMIN, src.mob, "toggles build mode [src.buildmode.is_active ? "off" : "on"]")
 
 	if(src.buildmode.is_active)
 		src.buildmode.deactivate()
@@ -228,19 +263,19 @@ ABSTRACT_TYPE(/datum/buildmode)
 		if (!usr.client.holder.buildmode_view)
 			usr.client.cmd_admin_aview()
 		usr.see_in_dark = initial(usr.see_in_dark)
-		usr.see_invisible = 16
+		usr.see_invisible = INVIS_SPOOKY
 	else
 		src.buildmode.activate()
 		if (!usr.client.holder.buildmode_view)
 			usr.client.cmd_admin_aview()
 		usr.see_in_dark = 10
-		usr.see_invisible = 21
+		usr.see_invisible = INVIS_ADVENTURE
 		src.show_popup_menus = 0
 
 /atom/movable/screen/buildmode/builddir
 	name = "Set direction"
 	density = 1
-	anchored = 1
+	anchored = ANCHORED
 	layer = HUD_LAYER + 1
 	plane = PLANE_HUD
 	dir = SOUTH
@@ -259,29 +294,29 @@ ABSTRACT_TYPE(/datum/buildmode)
 
 		if (icon_y <= 11)
 			if (icon_x <= 11)
-				dir = 10
+				dir = SOUTHWEST
 			else if (icon_x >= 22)
-				dir = 6
+				dir = SOUTHEAST
 			else
-				dir = 2
+				dir = SOUTH
 		else if (icon_y >= 22)
 			if (icon_x <= 11)
-				dir = 9
+				dir = NORTHWEST
 			else if (icon_x >= 22)
-				dir = 5
+				dir = NORTHEAST
 			else
-				dir = 1
+				dir = NORTH
 		else if (icon_x <= 16)
-			dir = 8
+			dir = WEST
 		else
-			dir = 4
+			dir = EAST
 
 		holder.dir = dir
 
 /atom/movable/screen/buildmode/buildhelp
 	name = "Click for help"
 	density = 1
-	anchored = 1
+	anchored = ANCHORED
 	layer = HUD_LAYER + 1
 	plane = PLANE_HUD
 	dir = NORTH
@@ -300,7 +335,7 @@ ABSTRACT_TYPE(/datum/buildmode)
 /atom/movable/screen/buildmode/buildquit
 	name = "Click to exit build mode"
 	density = 1
-	anchored = 1
+	anchored = ANCHORED
 	layer = HUD_LAYER + 1
 	plane = PLANE_HUD
 	dir = NORTH
@@ -319,7 +354,7 @@ ABSTRACT_TYPE(/datum/buildmode)
 /atom/movable/screen/buildmode/buildmode
 	name = "Click to select mode"
 	density = 1
-	anchored = 1
+	anchored = ANCHORED
 	layer = HUD_LAYER + 1
 	plane = PLANE_HUD
 	dir = NORTH
@@ -336,7 +371,7 @@ ABSTRACT_TYPE(/datum/buildmode)
 		if ("middle" in pa)
 			src.holder.add_to_hotkey_bar(src.holder.mode.copy())
 		else if ("left" in pa)
-			var/modename = input("Select new mode", "Select new mode", holder.mode.name) in sortList(holder.modes_cache)
+			var/modename = input("Select new mode", "Select new mode", holder.mode.name) in sortList(holder.modes_cache, /proc/cmp_text_asc)
 			if (modename == holder.mode.name && isnull(holder.mode.hotkey_number))
 				return
 			holder.select_mode(holder.modes_cache[modename])
@@ -345,7 +380,7 @@ ABSTRACT_TYPE(/datum/buildmode)
 
 /atom/movable/screen/buildmode/hotkey
 	density = 1
-	anchored = 1
+	anchored = ANCHORED
 	layer = HUD_LAYER + 1
 	plane = PLANE_HUD
 	dir = NORTH
@@ -368,13 +403,46 @@ ABSTRACT_TYPE(/datum/buildmode)
 		else if ("right" in pa)
 			mode.click_mode_right(pa.Find("ctrl"), pa.Find("alt"), pa.Find("shift"))
 
+	MouseWheel(delta_x, delta_y, location, control, params)
+		. = TRUE
+		var/current = 0
+		for(var/datum/buildmode/mode in holder.hotkey_bar)
+			if(mode == holder.mode)
+				current = mode.hotkey_number
+				break
+		if(delta_y > 0)
+			var/datum/buildmode/firstmode = null
+			for(var/datum/buildmode/mode in holder.hotkey_bar)
+				if(isnull(firstmode))
+					firstmode = mode
+				if(mode.hotkey_number > current)
+					holder.select_mode(mode)
+					return
+			holder.select_mode(firstmode)
+		else
+			var/datum/buildmode/lastmode = null
+			for(var/datum/buildmode/mode in holder.hotkey_bar)
+				if(mode.hotkey_number >= current)
+					break
+				lastmode = mode
+			if(isnull(lastmode))
+				for(var/datum/buildmode/mode in holder.hotkey_bar)
+					lastmode = mode
+			holder.select_mode(lastmode)
+
 
 var/image/buildmodeBlink = image('icons/effects/effects.dmi',"empdisable")//guH GUH GURGLE
+var/buildmodeBlinkCounter = 0
+var/buildmodeBlinkTick = 0
 /proc/blink(var/turf/T)
 	if (!T)
 		return
-
-	SPAWN_DBG(0)//WHY DOUBLE SPAWN AND NEW IMAGE EVERY BLINK IT MAKES SOMEPOTATO SAD
+	if(TIME != buildmodeBlinkTick)
+		if(buildmodeBlinkCounter++ > 100) return
+	else
+		buildmodeBlinkTick = TIME
+		buildmodeBlinkCounter = 0
+	SPAWN(0)//WHY DOUBLE SPAWN AND NEW IMAGE EVERY BLINK IT MAKES SOMEPOTATO SAD
 
 		T.overlays += buildmodeBlink
 		sleep(0.5 SECONDS)
